+++ sentry-ruby/spec/sentry/span_spec.rb
@@ -70,6 +70,16 @@ RSpec.describe Sentry::Span do
       expect(hash[:trace_id].length).to eq(32)
       expect(hash[:span_id].length).to eq(16)
     end
+
+    it 'has metric summary if present' do
+      key = [:c, 'incr', 'none', []]
+      subject.metrics_local_aggregator.add(key, 10)
+
+      hash = subject.to_hash
+      expect(hash[:_metrics_summary]).to eq({
+        'c:incr@none' => { count: 1, max: 10.0, min: 10.0, sum: 10.0, tags: {} }
+      })
+    end
   end
 
   describe "#to_sentry_trace" do